requirements.txt
Flask==3.0.3
pymongo==4.8.0
python-dotenv==1.0.1
flask-login==0.6.3